refactor: RedisService -> 두가지 역할을 하던 서비스클래스 분리
